cifsd: out internal FID representation uses 4 bytes

+++ glob.h
@@ -155,8 +155,8 @@ struct cifsd_work {
 	 * Current Local FID assigned compound response if SMB2 CREATE
 	 * command is present in compound request
 	 */
-	__u64				cur_local_fid;
-	__u64				cur_local_pfid;
+	unsigned int			cur_local_fid;
+	unsigned int			cur_local_pfid;
 	__u64				cur_local_sess_id;
 
 	int				state;
